TagとTaggedCharacterのtype guardをtagged-character.tsに移動
